Only kill the JVM when thread creation fails
Killing the JVM due to OOM will interrupt the heap dumper.

+++ jvmkill.c
@@ -28,10 +28,12 @@ resourceExhausted(
       const void *reserved,
       const char *description)
 {
-   fprintf(stderr,
-      "ResourceExhausted: %s: killing current process!\n", description);
-   kill(getpid(), SIGKILL);
-   _exit(1);
+   if (flags & JVMTI_RESOURCE_EXHAUSTED_THREADS) {
+      fprintf(stderr,
+         "ResourceExhausted: %s: killing current process!\n", description);
+      kill(getpid(), SIGKILL);
+      _exit(1);
+   }
 }
 
 JNIEXPORT jint JNICALL
